5_Agentic_chatbot/requirements.txt
langgraph
langchain-groq
python-dotenv
streamlit
pydantic